GarmentCorrectionNoteFacade.ReadById Parameter

+++ Com.DanLiris.Service.Purchasing.Lib/Interfaces/IGarmentCorrectionNoteFacade.cs
@@ -9,7 +9,7 @@ namespace Com.DanLiris.Service.Purchasing.Lib.Interfaces
     public interface IGarmentCorrectionNoteFacade
     {
         Tuple<List<GarmentCorrectionNote>, int, Dictionary<string, string>> Read(int Page = 1, int Size = 25, string Order = "{}", string Keyword = null, string Filter = "{}");
-        GarmentCorrectionNote ReadById(long id);
+        GarmentCorrectionNote ReadById(int id);
         Task<int> Create(GarmentCorrectionNote garmentCorrectionNote, string user, int clientTimeZoneOffset = 7);
     }
 }
